+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix(logging): fixed permissions on custom resource for when cloudwatch encryption is enabled in global-config
 - fix(s3): imported elb bucket policy attachment failed
 - fix(config/validation): make account email comparisons case insensitive
+- fix(config-service): only record global resources in home region
 
 ### Changed